integer sequence

+++ utils/utils_test.go
@@ -10,15 +10,15 @@ func TestProduct(t *testing.T) {
 	k := 3
 	i := 0
 
-	check := make([][]int, 8)
-	check[0] = []int{0, 0, 0}
-	check[1] = []int{0, 0, 1}
-	check[2] = []int{0, 1, 0}
-	check[3] = []int{0, 1, 1}
-	check[4] = []int{1, 0, 0}
-	check[5] = []int{1, 0, 1}
-	check[6] = []int{1, 1, 0}
-	check[7] = []int{1, 1, 1}
+	check := make([][]float64, 8)
+	check[0] = []float64{0, 0, 0}
+	check[1] = []float64{0, 0, 1}
+	check[2] = []float64{0, 1, 0}
+	check[3] = []float64{0, 1, 1}
+	check[4] = []float64{1, 0, 0}
+	check[5] = []float64{1, 0, 1}
+	check[6] = []float64{1, 1, 0}
+	check[7] = []float64{1, 1, 1}
 
 	for res := range Product(a, k) {
 		fmt.Println(res)
